+++ src/components/Home/__tests__/HomeComponent.test.js
@@ -1,7 +1,7 @@
 import React from "react";
 import { configure, shallow } from "enzyme";
 import Adapter from "enzyme-adapter-react-16";
-import HomeComponent from "./HomeComponent";
+import HomeComponent from "../HomeComponent";
 
 configure({ adapter: new Adapter() });
