+++ src/components/MusicStyleDetails/__tests__/MusicStyleDetails.test.js
@@ -2,7 +2,7 @@ import React from "react";
 import { configure, shallow } from "enzyme";
 import Adapter from "enzyme-adapter-react-16";
 import { BrowserRouter as Router, Route } from "react-router-dom";
-import MusicStyleContainer from "./MusicStyleContainer";
+import MusicStyleDetailsContainer from "../MusicStyleDetailsContainer";
 
 configure({ adapter: new Adapter() });
 
@@ -11,7 +11,7 @@ it("Renders MusicStyleContainer without crashing", () => {
   const paramsRouter = {
     paramsRouter: {
       match: {
-        params: { musicStyle: "jazz" }
+        params: { musicStyleDetail: "artists" }
       }
     }
   };
@@ -19,9 +19,9 @@ it("Renders MusicStyleContainer without crashing", () => {
     <Router>
       <Route
         exact
-        path="/:musicStyle"
+        path="/:musicStyle/:musicStyleDetail"
         render={() => (
-          <MusicStyleContainer
+          <MusicStyleDetailsContainer
             paramsRouter={paramsRouter}
             translateFunction={translateFunction}
           />
